Add files via upload
The code itself is the extension to the part of file.html extraction from the web page(in our case Moodle). Here one could see the part for titles and subtitles extraction (it is not necessary to be present in the code itself, but it is nice to see the relation of our sections right infant of the eyes), and the following filling the RDF graph with CourseSyllabus vocabulary.